// src/app/core/models/Hoghoghi/ModirShakhsHoghooghi.ts
export interface ModirShakhsHoghooghi {
    Id?: number;
    
    SematModirShakhsHoghooghiTypeName?: any;
    ShenasehMeli?: any;
    ShakhsHoghooghiName?: string;
    FirstName?: string;
    FamilyName?: string;
    ShakhsHoghooghiTypeId?: any;
    NationalCode?: any;

    ShakhsHoghooghiId?: number;
    ShakhsHaghighiId?: number;
    SematModirShakhsHoghooghiTypeId?: number;
    DarandehEmzaTahodavarShakhsHoghooghi?: any;
}